+++ public/idx.math.php
@@ -1,1 +1,30 @@
-This will be the math page.
+<form style="display: inline;" action="" method="POST">
+      Environemnt: <select name="environment" onchange="this.form.submit();">
+        <option value="develop">Development</option>
+        <option value="product">Production</option>
+        <option value="math" selected="">Math</option>
+      </select>
+</form>
+<br />
+This will be the math page.<br />
+<br />
+Favorites<br />
+  Percentage, Equations, Shapes, Bodies, Length, Speed, Temperature, Weight, 'Curreny Converter', Tip, 'Body Mass Index', Age, Unit Price<br />
+  <br />
+Algebra<br />
+  Percentage, Average, Proportion, Ratio, Equations, Fractions, GCF / LCM, Combinations, Prime Checker, Number Generator<br />
+  <br />
+Geometry<br />
+  Shapes, Bodies<br />
+  <br />
+Unit Converters<br />
+  Acceleration, Angle, Area, Cooking, Data Storage, Data Transfer, Energy, Force, Fuel, Length, Numeric Base, Power, Pressure, Ring Size, Roman Numerals, Shoe Size, Speed, Temperature, Time, Torque, Volume, Volumetric Flow, Weight, <br />
+  <br />
+Fiance<br />
+  Currency Converter, Unit Price, Sales Tax, Tip, Loan, Interest<br />
+  <br />
+Health<br />
+  Body Mass Index, Daily Caloric Burn, Body Fat Percentage<br />
+  <br />
+Misc<br />
+  Age, Date, Time, Mileage, Ohm's Law<br />