+++ src/App.js
@@ -1,0 +1,16 @@
+// src/App.js
+import React from 'react';
+import './App.css';
+import HelloWorld from './HelloWorld';
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <HelloWorld />
+      </header>
+    </div>
+  );
+}
+
+export default App;