+++ CHANGELOG.md
@@ -1,1 +1,22 @@
 CHANGELOG
+
+
+:: PHP
+
+Many php functions either produce a forward slash, indicating the end of folder/path
+
+  $path = dirname(__DIR__) . '/config/config.php'
+
+
+:: JavaScript
+
+JScript
+  tailwindcss-3.3.5.js
+Style (type="text/tailwindcss")
+
+
+- Non-indented/tab code means that it is included/required/imported
+
+Javascript libraries must be loaded in a particular order. If a library can not be loaded, "ace is not defined", then they are in the wrong order.
+
+jQuery, jQuery-ui, ace-editor, requirejs